Simplify the core API
